refactor:#2 내 정보 조회 위치 변경, AuthController -> MemberController
